API Remove MonologSyslogWriter, separate Logger test stub into own file
