+++ code/AuditFactory.php
@@ -2,6 +2,11 @@
 
 namespace SilverStripe\Auditor;
 
+use Exception;
+use Monolog\Formatter\LineFormatter;
+use Monolog\Handler\SyslogHandler;
+use Monolog\Logger;
+use Monolog\Processor\WebProcessor;
 use SilverStripe\Core\Injector\Factory;
 
 /**
@@ -10,30 +15,32 @@ use SilverStripe\Core\Injector\Factory;
  */
 class AuditFactory implements Factory
 {
-    public function create($service, array $params = array())
+    public function create($service, array $params = [])
     {
         if (!empty($params)) {
-            throw new \Exception('AuditFactory does not support passing params.');
+            throw new Exception('AuditFactory does not support passing params.');
         }
 
         $obj = null;
         switch ($service) {
             case 'AuditLogger':
-                $log = new \Monolog\Logger('audit');
-                $syslog = new \Monolog\Handler\SyslogHandler('SilverStripe_audit', LOG_AUTH, \Monolog\Logger::DEBUG);
-                $syslog->pushProcessor(new \Monolog\Processor\WebProcessor($_SERVER, array(
-                'url'         => 'REQUEST_URI',
-                'http_method' => 'REQUEST_METHOD',
-                'server'      => 'SERVER_NAME',
-                'referrer'    => 'HTTP_REFERER',
-                )));
-                    $syslog->pushProcessor(new RealIPProcessor());
-                    $formatter = new \Monolog\Formatter\LineFormatter("%level_name%: %message% %context% %extra%");
-                    $syslog->setFormatter($formatter);
-                    $log->pushHandler($syslog);
+                $log = new Logger('audit');
+                $syslog = new SyslogHandler('SilverStripe_audit', LOG_AUTH, Logger::DEBUG);
+                $syslog->pushProcessor(new WebProcessor($_SERVER, [
+                    'url'         => 'REQUEST_URI',
+                    'http_method' => 'REQUEST_METHOD',
+                    'server'      => 'SERVER_NAME',
+                    'referrer'    => 'HTTP_REFERER',
+                ]));
+
+                $syslog->pushProcessor(new RealIPProcessor());
+                $formatter = new LineFormatter("%level_name%: %message% %context% %extra%");
+                $syslog->setFormatter($formatter);
+                $log->pushHandler($syslog);
+
                 return $log;
             default:
-                throw new \Exception(sprintf("AuditFactory does not support creation of '%s'.", $service));
+                throw new Exception(sprintf("AuditFactory does not support creation of '%s'.", $service));
         }
     }
 }
